Bug-1943334: Taint node with NoSchedule effect when ovnkube pod is down
UseCase: During upgrades, when the ovnkube pod is deleted and
before/while it gets recreated if a new pod request comes, it
will fail.

When node networking is down, we should not allow pods to be created
on that node. This PR fixes this by adding a taint on the node with
the NoScedule effect when the ovnkube pod receives a SIGTERM and then
removes the taint when the ovnkube pod finishes setting up networking
and initializing the CNI config.

Signed-off-by: Surya Seetharaman <suryaseetharaman.9@gmail.com>